+++ src/modules/ce/components/defaultContacts/ProjectDefaultContactsCard.tsx
@@ -4,8 +4,8 @@ import SwimlaneLabel from './SwimlaneLabel';
 import CommonCard from '@/components/elements/CommonCard';
 import { CommonLabeledTextBlock } from '@/components/elements/CommonLabeledTextBlock';
 import CommonMenuButton from '@/components/elements/CommonMenuButton';
-import EditCeDefaultContactsModal from '@/modules/ce/components/admin/EditCeDefaultContactsModal';
 import DefaultContactNamesList from '@/modules/ce/components/defaultContacts/DefaultContactNamesList';
+import EditCeDefaultContactsModal from '@/modules/ce/components/defaultContacts/EditCeDefaultContactsModal';
 import { useRootPermissions } from '@/modules/permissions/useHasPermissionsHooks';
 import { cache } from '@/providers/apolloClient';
 import { ProjectWithCeDefaultContactsFragment } from '@/types/gqlTypes';
